Improved testing and also found an authenticate bug... Woot unit tests
