+++ tests/general.js
@@ -2,7 +2,6 @@
 const chai = require('chai');
 const chaiHttp = require('chai-http');
 const should = chai.should();
-
 chai.use(chaiHttp);
 
 module.exports = (test_url) => {
